+++ Resources/assets/js/helpers.js
@@ -1,0 +1,7 @@
+import encoding from './helpers/encoding';
+import form from './helpers/form';
+import security from './helpers/security';
+
+export {
+    encoding, form, security
+}
